Enhance project structure and testing framework; add modular arithmetic functions and tests

+++ include/number_theory/mod_inverse.h
@@ -6,12 +6,11 @@
 
 #pragma once
 
-#include <bits/stdc++.h>
 #include "../number_theory/binpow.h"
 
 using namespace std;
 
 // Fermat's inverse: a^(m-2) % m, valid when m is prime
-auto mod_inv_fermat = [](int64_t a, int64_t m) -> int64_t {
+auto mod_inv_fermat = [](int64_t a, int m) -> int64_t {
   return binpow_mod(a, m - 2, m);
 };
